refactor(models): refactor imports and typings in message transformation modules
- Consolidate and update import paths for types and utilities in anthropic-messages.ts, google-messages.ts, and prepare-request-body.ts
- Add explicit typing for some variables and parameters
- Remove redundant imports and improve code clarity
- Minor formatting improvements in transformAnthropicMessages function
